[ARCHIVES] Change archives attribute to be a file.
Archives is currently defined as a attr.string_list_dict which is a
dictionary whos values are a list of strings.
The versions structure is a dictionary of version to lists
of dictionaries.  There is no way to define this attribute structure
in bazel, so instead take a json file which contains the archives.